pyproject.toml
[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[project]
name = "kaufman"
version = "0.1.0"
description = "Barebones Streamlit web app"
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
authors = [
  { name = "Skyler Kaufman" },
  { name = "Ben Kaufman" }
]
dependencies = [
  "streamlit>=1.36.0",
  "pytesseract>=0.3.10",
  "pillow>=10.0.0",
  "opencv-python>=4.8.0",
  "pillow-heif>=0.15.0",
  "openai>=1.0.0"
]

[project.urls]
Homepage = "https://example.com"

[tool.hatch.build]
exclude = [
  ".git",
]

[tool.hatch.build.targets.wheel]
packages = ["kaufman"]
